requirements-test.txt
afinn==0.1
alembic==1.13.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
cytoolz==0.12.3
filelock==3.13.4
floret==0.10.5
fonttools==4.51.0
fsspec==2024.3.1
h11==0.14.0
huggingface-hub==0.22.2
hypothesis==6.100.1
idna==3.7
iniconfig==2.0.0
jellyfish==1.0.3
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
langcodes==3.3.0
lxml==5.2.1
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib==3.8.4
mistune==3.0.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.4.0
preshed==3.0.9
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
pyphen==0.14.0
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.19.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
sympy==1.12
textacy==0.13.0
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
wasabi==1.1.2
weasel==0.3.4
wordcloud==1.9.3
wsproto==1.2.0
